feat: US-04 - Split database/collection_ops.rs into modular ops files
Split the 650-line collection_ops.rs into focused modules:
- collection_ops.rs: dispatcher + shared ops (<200 NLOC)
- vector_ops.rs: vector collection create/get
- graph_ops.rs: graph collection create/get
- metadata_ops.rs: metadata-only collection create/get

Extracted register_vector_collection and register_graph_collection
helpers to deduplicate registry + observer + schema_version boilerplate.
